+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [catalog] FilesApp Survival Manual
 - [catalog] FilesApp The Life
 - [builder] `add_file()`'s `to` can now use `${APP_DIR:<ident>}` to target any app's home
+- [builder] Use reverse-proxy 1.5 with single-service fix
 
 ### Removed