+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.7.6] - 2024-02-05
+## [1.8.1] - 2024-02-05
+
+### Changed
+
+- [builder] Using metrics 0.3.0
+
+## [1.8.0] - 2024-01-30
 
 ### Added